+++ src/elements/account/createOrganization.tsx
@@ -3,7 +3,7 @@ import { useEffect } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
 import { useOutletContext } from 'react-router-dom'
 import { ClientsGrid } from '~components/Home/Clients'
-import { OrganizationCreate } from '~components/Organization/CreateOrg'
+import { OrganizationCreate } from '~components/Organization/Create'
 import { AuthOutletContextType } from '~elements/LayoutAuth'
 import { Routes } from '~src/router/routes'
